doParallel-package {doParallel}	R Documentation

The doParallel Package

Description

The doParallel package provides a parallel backend for the foreach/%dopar% function using the parallel package of R 2.14.0 and later.

Details

Further information is available in the following help topics:

registerDoParallel 	 register doParallel to be used by foreach/%dopar%

To see a tutorial introduction to the doParallel package, use vignette("gettingstartedParallel"). To see a tutorial introduction to the foreach package, use vignette("foreach").

To see a demo of doParallel computing the sinc function, use demo(sincParallel).

Some examples (in addition to those in the help pages) are included in the “examples” directory of the doParallel package. To list the files in the examples directory, use list.files(system.file("examples", package="doParallel")). To run the bootstrap example, use source(system.file("examples", "bootParallel.R", package="doParallel")). This is a simple benchmark, executing both sequentally and in parallel. There are many more examples that come with the foreach package, which will work with the doParallel package if it is registered as the parallel backend.

For a complete list of functions with individual help pages, use library(help="doParallel").

[Package doParallel version 1.0.17 Index]